requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
